fix: 🐛 permissions
